Fixes for the checkin
